fix: minor updates and changes and refactors
