+++ src/main/java/me/ninjamandalorian/implodusports/ImplodusPorts.java
@@ -2,16 +2,14 @@ package me.ninjamandalorian.implodusports;
 
 import com.palmergames.bukkit.towny.TownyAPI;
 import me.ninjamandalorian.implodusports.command.CommandHandler;
-import me.ninjamandalorian.implodusports.command.ImplodusPortsCommands;
 import me.ninjamandalorian.implodusports.data.DataManager;
 import me.ninjamandalorian.implodusports.data.PortDataManager;
-import me.ninjamandalorian.implodusports.dymap.DynmapHandler;
+import me.ninjamandalorian.implodusports.dynmap.DynmapHandler;
 import me.ninjamandalorian.implodusports.listener.*;
 import me.ninjamandalorian.implodusports.object.Port;
 import me.ninjamandalorian.implodusports.settings.Settings;
 import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import net.milkbowl.vault.economy.Economy;
-import org.checkerframework.checker.units.qual.C;
 import org.dynmap.DynmapAPI;
 import org.dynmap.markers.MarkerAPI;
 import org.bukkit.Bukkit;
