+++ src/main/java/me/ninjamandalorian/implodusports/listener/BlockListener.java
@@ -3,6 +3,7 @@ package me.ninjamandalorian.implodusports.listener;
 import me.ninjamandalorian.implodusports.object.Port;
 import net.md_5.bungee.api.ChatColor;
 import org.bukkit.block.Block;
+import org.bukkit.block.Sign;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -17,15 +18,16 @@ public class BlockListener implements Listener {
 
 		if (block.getType().toString().contains("SIGN")) {
 			Port port = Port.getPort(block.getLocation());
+			
+			Sign sign = (Sign) block.getState();
+			
 			if (port == null) return;
 
 			if (player.hasPermission("implodusports.admin.destroy")) {
 				Port.portDestroy(player, port);
-				return;
 			} else {
 				e.setCancelled(true);
 				player.sendMessage(ChatColor.translateAlternateColorCodes('&', "&9[&6iPorts&9] &cYou do not have permission to destroy this."));
-				return;
 			}
 		}
 	}
